+++ Sources/KeyboardKit/Layout/Providers/iPhoneKeyboardLayoutProvider.swift
@@ -72,4 +72,3 @@ open class iPhoneKeyboardLayoutProvider: BaseKeyboardLayoutProvider, KeyboardLay
         [.backspace]
     }
 }
-